Add files via upload
In the study “Spatial-temporal analysis of rainfall erosivity in southeastern brazil: an approach using machine learning algorithms.” we applied an approach to predict rainfall erosivity with Machine Learning Algorithms (ML) method.
The advantage of using ML in the study was the possibility of testing the predictions with a set of 122 covariates and with five algorithms (Random Forest-RF, Cubist-CUB, Support Vector Machine, Multivariate Adaptive Regression Spline, and Linear Model). Studies considering an extensive collection of covariates + different methods are scarce in the literature.
To test the algorithms, we selected a state of economic relevance and climatic diversity in Brazil:  The state is the second GDP of Brazil, highlighting its agricultural, being influenced by various climatic conditions. For testing the algorithms, we selected the Brazilian state of Minas Gerais.  The state is Brazil second GDP, highlighting its agriculture, and has a different climatic condition. Furthermore, regarding rainfall erosivity, there are previous studies by other prediction methods, allowing us to compare the results with ML predictions.